# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get short user details",
                "description": "Get short user details",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/ShortUserResource"
                    }
                }
            }
        },
        "/api/v1/users/{ID}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get full user details",
                "description": "Get full user details",
                "operationId": "details",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 11
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Vladimir Putin"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "http://b-ouquet.com/storage/avatars/3gpbiSkdpCbyIADiyb0wuopeSVnbT0gBCJfZi0GL_avatar.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "update-user",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Disable user",
                "description": "Disable user",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/users/{ID}/avatar": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Upload user avatar",
                "description": "Upload user avatar",
                "operationId": "upload-user-avatar",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "file",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ShortUserResource": {
                "properties": {
                    "id": {
                        "type": "int8",
                        "example": "53"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "http://b-ouquet.com/storage/avatars/3gpbiSkdpCbyIADiyb0wuopeSVnbT0gBCJfZi0GL_avatar.jpg"
                    },
                    "memorial": {
                        "type": "int8",
                        "example": "35"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShortUserResource"
                }
            }
        },
...
    }
}